[GStreamer][MediaStream] Move mock display source to its own cpp file https://bugs.webkit.org/show_bug.cgi?id=258072
Reviewed by Xabier Rodriguez-Calvar.

Moving code around for consistency purposes.

* Source/WebCore/platform/SourcesGStreamer.txt:
* Source/WebCore/platform/mediastream/gstreamer/MockDisplayCaptureSourceGStreamer.cpp: Copied from Source/WebCore/platform/mediastream/gstreamer/MockRealtimeVideoSourceGStreamer.cpp.
(WebCore::MockDisplayCaptureSourceGStreamer::create):
(WebCore::MockDisplayCaptureSourceGStreamer::MockDisplayCaptureSourceGStreamer):
(WebCore::MockDisplayCaptureSourceGStreamer::~MockDisplayCaptureSourceGStreamer):
(WebCore::MockDisplayCaptureSourceGStreamer::stopProducingData):
(WebCore::MockDisplayCaptureSourceGStreamer::requestToEnd):
(WebCore::MockDisplayCaptureSourceGStreamer::setMuted):
(WebCore::MockDisplayCaptureSourceGStreamer::videoFrameAvailable):
(WebCore::MockDisplayCaptureSourceGStreamer::capabilities):
(WebCore::MockDisplayCaptureSourceGStreamer::settings):
* Source/WebCore/platform/mediastream/gstreamer/MockDisplayCaptureSourceGStreamer.h: Copied from Source/WebCore/platform/mediastream/gstreamer/MockRealtimeVideoSourceGStreamer.h.
* Source/WebCore/platform/mediastream/gstreamer/MockRealtimeVideoSourceGStreamer.cpp:
(WebCore::MockDisplayCaptureSourceGStreamer::create): Deleted.
(WebCore::MockDisplayCaptureSourceGStreamer::MockDisplayCaptureSourceGStreamer): Deleted.
(WebCore::MockDisplayCaptureSourceGStreamer::~MockDisplayCaptureSourceGStreamer): Deleted.
(WebCore::MockDisplayCaptureSourceGStreamer::stopProducingData): Deleted.
(WebCore::MockDisplayCaptureSourceGStreamer::requestToEnd): Deleted.
(WebCore::MockDisplayCaptureSourceGStreamer::setMuted): Deleted.
(WebCore::MockDisplayCaptureSourceGStreamer::videoFrameAvailable): Deleted.
(WebCore::MockDisplayCaptureSourceGStreamer::capabilities): Deleted.
(WebCore::MockDisplayCaptureSourceGStreamer::settings): Deleted.
* Source/WebCore/platform/mediastream/gstreamer/MockRealtimeVideoSourceGStreamer.h:
* Source/WebCore/platform/mock/MockRealtimeMediaSourceCenter.cpp:

Canonical link: https://commits.webkit.org/265316@main
