+++ Source/WebCore/platform/mediastream/gstreamer/MockRealtimeVideoSourceGStreamer.cpp
@@ -52,101 +52,6 @@ CaptureSourceOrError MockRealtimeVideoSource::create(String&& deviceID, AtomStri
     return source;
 }
 
-CaptureSourceOrError MockDisplayCaptureSourceGStreamer::create(const CaptureDevice& device, MediaDeviceHashSalts&& hashSalts, const MediaConstraints* constraints, PageIdentifier pageIdentifier)
-{
-    auto mockSource = adoptRef(*new MockRealtimeVideoSourceGStreamer(String { device.persistentId() }, AtomString { device.label() }, MediaDeviceHashSalts { hashSalts }));
-
-    if (constraints) {
-        if (auto error = mockSource->applyConstraints(*constraints))
-            return WTFMove(error->message);
-    }
-
-    Ref<RealtimeMediaSource> source = adoptRef(*new MockDisplayCaptureSourceGStreamer(device, WTFMove(mockSource), WTFMove(hashSalts), pageIdentifier));
-    return source;
-}
-
-MockDisplayCaptureSourceGStreamer::MockDisplayCaptureSourceGStreamer(const CaptureDevice& device, Ref<MockRealtimeVideoSourceGStreamer>&& source, MediaDeviceHashSalts&& hashSalts, PageIdentifier pageIdentifier)
-    : RealtimeVideoCaptureSource(device, WTFMove(hashSalts), pageIdentifier)
-    , m_source(WTFMove(source))
-    , m_deviceType(device.type())
-{
-    m_source->addVideoFrameObserver(*this);
-}
-
-MockDisplayCaptureSourceGStreamer::~MockDisplayCaptureSourceGStreamer()
-{
-    m_source->removeVideoFrameObserver(*this);
-}
-
-void MockDisplayCaptureSourceGStreamer::stopProducingData()
-{
-    m_source->removeVideoFrameObserver(*this);
-    m_source->stop();
-}
-
-void MockDisplayCaptureSourceGStreamer::requestToEnd(Observer& callingObserver)
-{
-    RealtimeMediaSource::requestToEnd(callingObserver);
-    m_source->removeVideoFrameObserver(*this);
-    m_source->requestToEnd(callingObserver);
-}
-
-void MockDisplayCaptureSourceGStreamer::setMuted(bool isMuted)
-{
-    RealtimeMediaSource::setMuted(isMuted);
-    m_source->setMuted(isMuted);
-}
-
-void MockDisplayCaptureSourceGStreamer::videoFrameAvailable(VideoFrame& videoFrame, VideoFrameTimeMetadata metadata)
-{
-    RealtimeMediaSource::videoFrameAvailable(videoFrame, metadata);
-}
-
-const RealtimeMediaSourceCapabilities& MockDisplayCaptureSourceGStreamer::capabilities()
-{
-    if (!m_capabilities) {
-        RealtimeMediaSourceCapabilities capabilities(settings().supportedConstraints());
-
-        // FIXME: what should these be?
-        // Currently mimicking the values for SCREEN-1 in MockRealtimeMediaSourceCenter.cpp::defaultDevices()
-        capabilities.setWidth(CapabilityValueOrRange(1, 1920));
-        capabilities.setHeight(CapabilityValueOrRange(1, 1080));
-        capabilities.setFrameRate(CapabilityValueOrRange(.01, 30.0));
-
-        m_capabilities = WTFMove(capabilities);
-    }
-    return m_capabilities.value();
-}
-
-const RealtimeMediaSourceSettings& MockDisplayCaptureSourceGStreamer::settings()
-{
-    if (!m_currentSettings) {
-        RealtimeMediaSourceSettings settings;
-        settings.setFrameRate(frameRate());
-
-        m_source->ensureIntrinsicSizeMaintainsAspectRatio();
-        auto size = m_source->size();
-        settings.setWidth(size.width());
-        settings.setHeight(size.height());
-        settings.setDeviceId(hashedId());
-        settings.setDisplaySurface(m_source->mockScreen() ? DisplaySurfaceType::Monitor : DisplaySurfaceType::Window);
-        settings.setLogicalSurface(false);
-
-        RealtimeMediaSourceSupportedConstraints supportedConstraints;
-        supportedConstraints.setSupportsFrameRate(true);
-        supportedConstraints.setSupportsWidth(true);
-        supportedConstraints.setSupportsHeight(true);
-        supportedConstraints.setSupportsDisplaySurface(true);
-        supportedConstraints.setSupportsLogicalSurface(true);
-        supportedConstraints.setSupportsDeviceId(true);
-
-        settings.setSupportedConstraints(supportedConstraints);
-
-        m_currentSettings = WTFMove(settings);
-    }
-    return m_currentSettings.value();
-}
-
 MockRealtimeVideoSourceGStreamer::MockRealtimeVideoSourceGStreamer(String&& deviceID, AtomString&& name, MediaDeviceHashSalts&& hashSalts)
     : MockRealtimeVideoSource(WTFMove(deviceID), WTFMove(name), WTFMove(hashSalts), { })
 {
