+++ Source/WebCore/platform/mediastream/gstreamer/MockRealtimeVideoSourceGStreamer.h
@@ -44,45 +44,6 @@ private:
     std::unique_ptr<GStreamerVideoCapturer> m_capturer;
 };
 
-class MockDisplayCaptureSourceGStreamer : public RealtimeVideoCaptureSource, RealtimeMediaSource::VideoFrameObserver {
-public:
-    static CaptureSourceOrError create(const CaptureDevice&, MediaDeviceHashSalts&&, const MediaConstraints*, PageIdentifier);
-
-    void requestToEnd(Observer&) final;
-    bool isProducingData() const final { return m_source->isProducingData(); }
-    void setMuted(bool isMuted) final;
-    const IntSize size() const final { return m_source->size(); }
-
-#if !RELEASE_LOG_DISABLED
-    const char* logClassName() const final { return "MockDisplayCaptureSourceGStreamer"; }
-#endif
-
-protected:
-    // RealtimeMediaSource::VideoFrameObserver
-    void videoFrameAvailable(VideoFrame&, VideoFrameTimeMetadata) final;
-
-    void generatePresets() override { };
-    const Vector<VideoPreset>& presets() final { return m_presets; }
-
-private:
-    MockDisplayCaptureSourceGStreamer(const CaptureDevice&, Ref<MockRealtimeVideoSourceGStreamer>&&, MediaDeviceHashSalts&&, PageIdentifier);
-    ~MockDisplayCaptureSourceGStreamer();
-
-    void startProducingData() final { m_source->start(); }
-    void stopProducingData() final;
-    void settingsDidChange(OptionSet<RealtimeMediaSourceSettings::Flag>) final { m_currentSettings = { }; }
-    bool isCaptureSource() const final { return true; }
-    const RealtimeMediaSourceCapabilities& capabilities() final;
-    const RealtimeMediaSourceSettings& settings() final;
-    CaptureDevice::DeviceType deviceType() const final { return m_deviceType; }
-
-    Vector<VideoPreset> m_presets;
-    Ref<MockRealtimeVideoSourceGStreamer> m_source;
-    CaptureDevice::DeviceType m_deviceType;
-    std::optional<RealtimeMediaSourceCapabilities> m_capabilities;
-    std::optional<RealtimeMediaSourceSettings> m_currentSettings;
-};
-
 } // namespace WebCore
 
 #endif // ENABLE(MEDIA_STREAM) && USE(GSTREAMER)
